# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AllocationOut": {
        "properties": {
          "artist_id": {
            "title": "Artist Id",
            "type": "integer"
          },
          "line_detail": {
            "additionalProperties": true,
            "title": "Line Detail",
            "type": "object"
          },
          "net_payable": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Payable",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "stage_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Name"
          }
        },
        "required": [
          "artist_id",
          "stage_name",
          "period",
          "net_payable",
          "line_detail"
        ],
        "title": "AllocationOut",
        "type": "object"
      },
...
      "BatchDetailOut": {
        "properties": {
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/AllocationOut"
            },
            "title": "Allocations",
            "type": "array"
          },
          "batch": {
            "$ref": "#/components/schemas/BatchOut"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/FlagOut"
            },
            "title": "Flags",
            "type": "array"
          },
          "promotion": {
            "$ref": "#/components/schemas/PromotionPreviewOut"
          }
        },
        "required": [
          "batch",
          "allocations",
          "flags",
          "promotion"
        ],
        "title": "BatchDetailOut",
        "type": "object"
      },
      "BatchOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "n_allocations": {
            "default": 0,
            "title": "N Allocations",
            "type": "integer"
          },
          "n_flags": {
            "default": 0,
            "title": "N Flags",
            "type": "integer"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "status": {
            "enum": [
              "proposed",
              "approved",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_by_run": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Run"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "total_net_payable": {
            "default": "0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Net Payable",
            "type": "string"
          }
        },
        "required": [
          "id",
          "period",
          "status",
          "submitted_by_run",
          "summary",
          "created_at"
        ],
        "title": "BatchOut",
        "type": "object"
      },
...
      "FlagOut": {
        "properties": {
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "severity",
          "payload"
        ],
        "title": "FlagOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromotionPreviewOut": {
        "description": "The diff-style \"what changes if approved\" panel.",
        "properties": {
          "allocation_total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Allocation Total",
            "type": "string"
          },
          "n_paid_artists": {
            "title": "N Paid Artists",
            "type": "integer"
          },
          "n_staged_lines": {
            "title": "N Staged Lines",
            "type": "integer"
          },
          "staged_gross_by_currency": {
            "additionalProperties": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "title": "Staged Gross By Currency",
            "type": "object"
          },
          "statements_to_promote": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Statements To Promote",
            "type": "array"
          }
        },
        "required": [
          "statements_to_promote",
          "n_staged_lines",
          "staged_gross_by_currency",
          "allocation_total",
          "n_paid_artists"
        ],
        "title": "PromotionPreviewOut",
        "type": "object"
      },
      "RejectIn": {
        "properties": {
          "note": {
            "description": "rejection requires a note",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "RejectIn",
        "type": "object"
      },
...
      "ReviewActionIn": {
        "properties": {
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          }
        },
        "title": "ReviewActionIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agent platform for music label operations — contracts, catalog, royalty statements. Agents propose; humans approve. Everything is traced.",
    "title": "Backline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/review/batches": {
      "get": {
        "operationId": "list_batches_review_batches_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "proposed",
              "enum": [
                "proposed",
                "approved",
                "rejected",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchOut"
                  },
                  "title": "Response List Batches Review Batches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batches",
        "tags": [
          "review"
        ]
      }
    },
    "/review/batches/{batch_id}": {
      "get": {
        "operationId": "get_batch_review_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch",
        "tags": [
          "review"
        ]
      }
    },
    "/review/batches/{batch_id}/approve": {
      "post": {
        "operationId": "approve_batch_review_batches__batch_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Batch",
        "tags": [
          "review"
        ]
      }
    },
    "/review/batches/{batch_id}/reject": {
      "post": {
        "description": "Reject a proposed batch. A note is required — 'no' with no reason is not a\nreview.",
        "operationId": "reject_batch_review_batches__batch_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Batch",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}